+++ src/examples/java/com/nylas/examples/other/EventsExample.java
@@ -62,6 +62,9 @@ public class EventsExample {
 	protected static void basicEventCrud(Events events, Calendar primary) throws IOException, RequestFailedException {
 		Event created = createBasicEvent(events, primary);
 		System.out.println("Created: " + created);
+
+		Participant partier = new Participant("hamilton@example.com");
+		partier.name("Alexander Hamilton");
 		
 		Participant partier1 = new Participant("hmulligan@example.com");
 		partier1.name("Hercules Mulligan");
@@ -80,15 +83,14 @@ public class EventsExample {
 		created.setTitle("Nonsurprise Party");
 		created.setBusy(false);
 		created.setLocation("Lake Merritt");
-		created.setParticipants(Arrays.asList(created.getParticipants().get(0), partier1, partier2, partier3));
-		created.setRecurrence(new Recurrence(startTz.getId(), Arrays.asList("RRULE:FREQ=WEEKLY;BYDAY=TH")));
+		created.setRecurrence(new Recurrence(startTz.getId(), Collections.singletonList("RRULE:FREQ=WEEKLY;BYDAY=TH")));
 
 		Map<String, String> metadata = new HashMap<>();
 		metadata.put("event_category", "gathering");
 		created.setMetadata(metadata);
 
 		Event.Conferencing conferencing = new Event.Conferencing();
-		conferencing.setProvider("Zoom Meeting");
+		conferencing.setProvider(Event.Conferencing.ConferencingProviders.ZOOM);
 		Event.Conferencing.Details details = new Event.Conferencing.Details();
 		details.setMeetingCode("213");
 		details.setPassword("xyz");
@@ -99,8 +101,18 @@ public class EventsExample {
 
 		Event updated = events.update(created, true);
 		System.out.println("Updated: " + updated);
+
+		Event.EmailNotification notification = new Event.EmailNotification();
+		notification.setMinutesBeforeEvent(60);
+		notification.setSubject("Test Event Notification");
+		notification.setBody("Reminding you about our meeting.");
+		updated.setNotifications(Collections.singletonList(notification));
+		updated.setParticipants(Arrays.asList(partier, partier1, partier2, partier3));
+
+		updated = events.update(updated, true);
+		System.out.println("Updated: " + updated);
 		
-		events.delete(created.getId(), true);
+		events.delete(updated.getId(), true);
 		System.out.println("Deleted");
 	}
 
@@ -116,7 +128,7 @@ public class EventsExample {
 		Event updated = events.update(created, true);
 		System.out.println("Updated: " + updated);
 
-		events.delete(created.getId(), true);
+		events.delete(updated.getId(), true);
 		System.out.println("Deleted");
 	}
 
@@ -152,8 +164,6 @@ public class EventsExample {
 
 		Event event = new Event(primary.getId(), new Timespan(startTime, endTime));
 		event.setTitle("Surprise Party");
-		Participant partier = new Participant("hamilton@example.com");
-		partier.name("Alexander Hamilton");
 
 		Event created = events.create(event, true);
 		System.out.println("Created: " + created);