+++ src/main/java/com/nylas/JsonHelper.java
@@ -19,6 +19,7 @@ public class JsonHelper {
 	static {
 		moshi = new Moshi.Builder()
 				.add(Event.WHEN_JSON_FACTORY)
+				.add(Event.EVENT_NOTIFICATION_JSON_FACTORY)
 				.build();
 	}